Fixed an invalid login bug
This one was interesting. I'm not totally sure why it was occuring but it was due to dismissing the loading animation (LoadingController). On google I found lots of interesting other cases of problems with this LoadingController and other similar components. It seemed that it would only work when inside the .then() of the signInWithEmailAndPassword and I sense that this was somthing to do with the fact that the code before it is dismissed changes the page that we are on. So maybe something to do with presenting a loadingcontroller, navigating to a page, and then dismissing it? I also discovered that it is best practice to place navigation code after a dismiss() inside a .then() like so, loading.dismiss().then(() => this.navCtrl.setRoot(Page));.